final mods

+++ tooling/prod/log.c
@@ -13,7 +13,7 @@ void log_init()
 void log_to_stream_(FILE *stream, const char *src, const char *file, const int line, const char *fmt, ...)
 {
     va_list args;
-    float curr_time = ((float)clock() - begin) / CLOCKS_PER_SEC;
+    double curr_time = ((double)clock() - begin) / CLOCKS_PER_SEC;
     
     fprintf(stream, "[ %08.2f ] [%s %s:%d]: ", curr_time, src, file, line);
